fixed a bug about the disappear year of an extincted species. and add some R scripts to analysis the results.

+++ src/EES3D/Species.cpp
@@ -97,7 +97,7 @@ vector<double> Species::getNicheBreadthEvolutionRatio(){
 int Species::getNicheBreadthEvolutionParentLevel(){
     return nicheBreadthEvolutionParentLevel;
 }
-int Species::getNicheEnvolutionIndividualRatio(){
+double Species::getNicheEnvolutionIndividualRatio(){
     return nicheEnvolutionIndividualRatio;
 }
 double Species::getNicheBreadthEvolutionRandomRange(){
